requirements.txt
# ── Core Framework ──────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ── ML & Vision ─────────────────────────────────────────────────
ultralytics>=8.2.18          # YOLOv8 (auto-selects Python 3.13 compatible wheel)
opencv-python-headless>=4.10.0
Pillow>=10.4.0               # 10.4+ has Python 3.13 wheels
numpy>=1.26.4
scikit-learn>=1.4.2
lightgbm>=4.3.0

# ── Geo & Spatial ────────────────────────────────────────────────
geopandas>=0.14.4
shapely>=2.0.4
requests>=2.31.0

# ── Data & Utilities ────────────────────────────────────────────
pandas>=2.2.2
python-dotenv>=1.0.1
joblib>=1.4.2
httpx>=0.27.0

# ── Testing ─────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.6
